# Improve performance of Context.IsCyclical(...)