--- a/src/Ninject/Activation/Context.cs
+++ b/src/Ninject/Activation/Context.cs
@@ -31,6 +31,7 @@ namespace Ninject.Activation
     using Ninject.Parameters;
     using Ninject.Planning;
     using Ninject.Planning.Bindings;
+    using Ninject.Planning.Targets;
 
     /// <summary>
     /// Contains information about the activation of a single instance.
@@ -160,7 +161,7 @@ namespace Ninject.Activation
         public object Resolve()
         {
             if (this.Request.ActiveBindings.Contains(this.Binding) &&
-                this.IsCyclical(this.Request.ParentRequest))
+                IsCyclical(this.Request.ParentRequest, this.Request.Target))
             {
                 throw new ActivationException(ExceptionFormatter.CyclicalDependenciesDetected(this));
             }
@@ -185,6 +186,21 @@ namespace Ninject.Activation
             {
                 this.cachedScope = null;
             }
+        }
+
+        private static bool IsCyclical(IRequest request, ITarget target)
+        {
+            if (request == null)
+            {
+                return false;
+            }
+
+            if (request.Target == target)
+            {
+                return true;
+            }
+
+            return IsCyclical(request.ParentRequest, target);
         }
 
         private object ResolveInternal(object scope)
@@ -242,21 +258,6 @@ namespace Ninject.Activation
             }
 
             return reference.Instance;
-        }
-
-        private bool IsCyclical(IRequest request)
-        {
-            if (request == null)
-            {
-                return false;
-            }
-
-            if (request.Target == this.Request.Target)
-            {
-                return true;
-            }
-
-            return this.IsCyclical(request.ParentRequest);
         }
     }
 }