* Changed `Context.IsCyclical(...)` to a static method.
* Add **ReadOnlyKernel** benchmark for complex service.

**Before:**

|                        Method |     Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
------------------------------ |---------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
 GetOfT_Complex_Singlethreaded | 34.57 us | 0.6664 us | 0.6843 us |      6.7139 |           - |           - |             28400 B |
  GetOfT_Complex_Multithreaded | 12.47 us | 0.1880 us | 0.1758 us |      6.8500 |           - |           - |                 1 B |
**After:**

|                       Method |     Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
------------------------------ |---------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
 GetOfT_Complex_Singlethreaded | 33.47 us | 0.3283 us | 0.2910 us |      6.7139 |           - |           - |             28400 B |
  GetOfT_Complex_Multithreaded | 12.38 us | 0.1152 us | 0.1078 us |      6.8500 |           - |           - |                 1 B |